Start work on new score format

+++ src/App/Osu/HUD.h
@@ -87,8 +87,8 @@ class HUD : public OsuScreen {
     std::vector<ScoreboardSlot *> slots;
     MD5Hash beatmap_md5;
 
-    f64 live_pp = 0.0;
-    f64 live_stars = 0.0;
+    f32 live_pp = 0.0;
+    f32 live_stars = 0.0;
 
    private:
     struct CURSORTRAIL {
