+++ src/App/Osu/pp.cpp
@@ -21,7 +21,7 @@ gradual_pp* init_gradual_pp(DatabaseBeatmap* diff, u32 mod_flags, f32 ar, f32 cs
 }
 
 gradual_pp* calculate_gradual_pp(gradual_pp* pp, i32 cur_hitobject, i32 max_combo, i32 num_300, i32 num_100, i32 num_50,
-                                 i32 num_misses, f64* stars_out, f64* pp_out) {
+                                 i32 num_misses, f32* stars_out, f32* pp_out) {
     Timer t;
     t.start();
